@@ -18,10 +18,10 @@
     <properties>
         <mainClass>io.helidon.microprofile.cdi.Main</mainClass>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <org.jboss.jandex.version>2.1.1.Final</org.jboss.jandex.version>
-        <helidon.version>2.4.0</helidon.version>
+        <helidon.version>3.0.0</helidon.version>
         <mockito.core.version>3.12.4</mockito.core.version>
     </properties>
 
@@ -250,6 +190,15 @@
 
     <build>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>17</source>
+                    <target>17</target>
+                    <compilerArgs>--enable-preview</compilerArgs>
+                </configuration>
+            </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-dependency-plugin</artifactId>
