@@ -8,7 +8,7 @@ import com.neo.r2.ts.impl.map.scaling.MapScalingService;
 import com.neo.r2.ts.impl.persistence.entity.Heatmap;
 import com.neo.r2.ts.impl.persistence.entity.HeatmapType;
 import com.neo.r2.ts.impl.persistence.entity.Match;
-import com.neo.r2.ts.impl.persistence.searchable.MatchEvent;
+import com.neo.r2.ts.impl.persistence.searchable.MatchEventSearchable;
 import com.neo.util.common.impl.exception.InternalJsonException;
 import com.neo.util.common.impl.exception.InternalLogicException;
 import com.neo.util.common.impl.json.JsonUtil;
@@ -25,9 +25,9 @@ import com.neo.util.framework.api.persistence.search.SearchResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.transaction.RollbackException;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.transaction.RollbackException;
 import java.util.*;
 
 @ApplicationScoped
@@ -35,7 +35,7 @@ public class HeatmapGeneratorImpl {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(HeatmapGeneratorImpl.class);
 
-    protected static final SearchAggregation COUNT_AGGREGATION = new SimpleFieldAggregation("count", MatchEvent.F_MATCH_ID, SearchAggregation.AggregationType.COUNT);
+    protected static final SearchAggregation COUNT_AGGREGATION = new SimpleFieldAggregation("count", MatchEventSearchable.F_MATCH_ID, SearchAggregation.AggregationType.COUNT);
 
     protected static final String PLAYER_POS_X = "entity.position.x";
     protected static final String PLAYER_POS_Y = "entity.position.y";
@@ -52,7 +52,10 @@ public class HeatmapGeneratorImpl {
     MapScalingService mapScalingService;
 
     public Heatmap calculateMap(String map) {
-        Heatmap heatmap = calculate(mapScalingService.getMap(map), List.of(new ExplicitSearchCriteria("map", map)), PIXELS_PER_CALL);
+        Heatmap heatmap = calculate(
+                mapScalingService.getMap(map).orElseThrow(() -> new InternalLogicException("Map not supported")),
+                List.of(new ExplicitSearchCriteria("map", map)),
+                PIXELS_PER_CALL);
         heatmap.setType(HeatmapType.FULL_MAP_AGGREGATION);
         heatmap.setDescription("This is an autogenerated heatmap for the map " + map);
         try {
@@ -70,7 +73,8 @@ public class HeatmapGeneratorImpl {
             List<SearchCriteria> searchCriteriaList = new ArrayList<>();
             switch (heatmapType) {
                 case PLAYER_POSITION:
-                    searchCriteriaList.add(new ExplicitSearchCriteria(MatchEvent.F_MATCH_ID, match.get().getId().toString()));
+                    searchCriteriaList.add(new ExplicitSearchCriteria(
+                            MatchEventSearchable.F_MATCH_ID, match.get().getId().toString()));
                     break;
                 case FULL_MAP_AGGREGATION:
                 case CUSTOM:
@@ -78,7 +82,10 @@ public class HeatmapGeneratorImpl {
                     throw new InternalLogicException("Unsupported heatmap type");
             }
 
-            Heatmap heatmap = calculate(mapScalingService.getMap(match.get().getMap()),searchCriteriaList, PIXELS_PER_CALL);
+            Heatmap heatmap = calculate(
+                    mapScalingService.getMap(match.get().getMap()).orElseThrow(() -> new InternalLogicException("Map not supported")),
+                    searchCriteriaList,
+                    PIXELS_PER_CALL);
             heatmap.setType(heatmapType);
             heatmap.setDescription("This is an autogenerated heatmap for the match " + matchId);
             try {
@@ -97,9 +104,9 @@ public class HeatmapGeneratorImpl {
             Heatmap heatmap = new Heatmap();
             heatmap.setType(HeatmapType.CUSTOM);
             heatmap.setHighestCount(0L);
-            heatmap.setMap(gameMap.getName());
+            heatmap.setMap(gameMap.name());
 
-            MapScale mapScale = gameMap.getMapScale();
+            MapScale mapScale = gameMap.scale();
             ObjectNode result =  JsonUtil.emptyObjectNode();
             ArrayNode resultArray = JsonUtil.emptyArrayNode();
             result.set("entries", resultArray);
