@@ -1,19 +1,19 @@
-package com.neo.r2.ts.impl.persistence;
+package com.neo.r2.ts.impl.match;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-import javax.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.context.ApplicationScoped;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
 @ApplicationScoped
-public class GlobalGameState {
+public class GlobalMatchState {
 
     protected Map<UUID, JsonNode> currentMatchState = new HashMap<>();
 
-    public void setCurrentMatchState(JsonNode state) {
-        currentMatchState.put(UUID.fromString(state.get("matchId").asText()), state);
+    public void setCurrentMatchState(String matchId, JsonNode state) {
+        currentMatchState.put(UUID.fromString(matchId), state);
     }
 
     public JsonNode getCurrentMatchState(UUID uuid) {
